requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3

numpy==1.26.4
pandas==1.5.3
matplotlib==3.7.3
scipy==1.10.1
scikit-learn==1.2.2
joblib==1.3.2
threadpoolctl==3.1.0

tensorflow-intel==2.16.1
keras>=3.0.0
tensorboard==2.16.1
protobuf==3.20.3
h5py==3.10.0
absl-py==1.4.0
gast==0.4.0
google-pasta==0.2.0
opt-einsum==3.3.0
wrapt==1.14.1
termcolor==2.3.0
flatbuffers==23.5.26
grpcio==1.54.0
libclang==16.0.6

requests==2.31.0
urllib3==2.0.7
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4

python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
packaging==23.2
reportlab==3.6.12
